% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap_instrument.R
\name{redcap_instrument_server}
\alias{redcap_instrument_server}
\title{REDCap Instrument Server}
\usage{
redcap_instrument_server(id, redcap_vars, subject_id)
}
\arguments{
\item{id}{The module namespace}

\item{redcap_vars}{Variables passed from the REDCap Setup module}

\item{subject_id}{The currently selected subject id}
}
\value{
REDCap records from the currently selected project
}
\description{
REDCap Instrument Server
}
