% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap_setup.R
\name{redcap_setup_server}
\alias{redcap_setup_server}
\title{REDCap Setup Server}
\usage{
redcap_setup_server(id, reset)
}
\arguments{
\item{id}{The module namespace}

\item{reset}{If a problem with the instrument is encountered and acknowledged, reset the configuration}
}
\value{
REDCap connection variables and project information
}
\description{
REDCap Setup Server
}
