% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\name{safe_exportRecords}
\alias{safe_exportRecords}
\title{REDCap Safe Export Records}
\usage{
safe_exportRecords(rc_con, rc_field_names)
}
\arguments{
\item{rc_con}{A REDCap API Connection Object}

\item{rc_field_names}{The field names for a REDCap instrument}
}
\value{
A data frame containing existing REDCap records, or an empty data
frame with the structure of what the records would look like.
}
\description{
A safe wrapper around \link[redcapAPI:recordsMethods]{redcapAPI::exportRecords} that does not fail when records
are requested from an empty REDCap project. In the event of an empty project,
field names are used to create an empty data structure.
}
\keyword{internal}
