% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap_setup.R
\name{redcap_setup_server}
\alias{redcap_setup_server}
\title{REDCap Setup Server}
\usage{
redcap_setup_server(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
REDCap connection variables and project information
}
\description{
REDCap Setup Server
}
