% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{package}
\name{shinyREDCap}
\alias{shinyREDCap}
\title{shinyREDCap: Access REDCap Instruments through R Shiny}
\description{
A shiny module to connect you to a REDCap project and perform a chart review from within an R Shiny Application.
}
